# Limit GroupByCombineOperator to use 2 * numCores threads instead of creating one task per operator